moved BlackJack
